Refactor code to allow custom serverList
Before the code for loading and filtering the serverList was in the `query` function, this meant you couldn't override the server list in a simple way. Now there is a `prepareList` function that loads and filters the query list seperately to `query`.

The advantage of this is you can provide a custom serverList just by setting `dnsyo.lookup.serverList` on a class instance.

Woo flexibility.